backend/requirements.txt
Authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary
gunicorn
fastapi 
uvicorn 
python-multipart 
aiofiles
httpx
python-jose[cryptography]==3.3.0
pyjwt[crypto]
passlib[bcrypt]
sqlalchemy
itsdangerous
pydantic
pydantic[email]